Fix platform issues and modernize plugin structure
- add Swift Package Manager support for biometric_storage_darwin (#144)

- map Darwin errSecAuthFailed to AuthException biometricsChanged (#128)

- relax Android plugin Java target and modernize example Android config (#140)

- fall back when StrongBox-backed key generation fails on Android 15 devices (#127)

- replace direct win32 dependency with local Windows credential FFI bindings

- update web/example integration and related platform validation artifacts
