+++ .gitignore
@@ -40,6 +40,8 @@ android/.idea
 .flutter-plugins-dependencies
 example/android/app/_debug-full-r8-config.txt
 
+.build/
+.swiftpm/
 
 .vscode
