+++ biometric_storage/example/android/gradle.properties
@@ -1,6 +1,6 @@
 org.gradle.jvmargs=-Xmx1536M
 
 android.useAndroidX=true
-android.defaults.buildfeatures.buildconfig=true
 android.nonTransitiveRClass=false
 android.nonFinalResIds=false
+android.suppressUnsupportedCompileSdk=36
