+++ biometric_storage/example/android/settings.gradle
@@ -19,7 +19,7 @@ pluginManagement {
 plugins {
     id "dev.flutter.flutter-plugin-loader" version "1.0.0" // apply true
     id "com.android.application" version "8.8.0" apply false
-    id "org.jetbrains.kotlin.android" version "1.8.10" apply false
+    id "org.jetbrains.kotlin.android" version "2.1.0" apply false
 }
 
 include ":app"
