+++ biometric_storage_android/android/build.gradle
@@ -23,23 +23,18 @@ rootProject.allprojects {
 
 apply plugin: 'com.android.library'
 apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-kapt'
-
-kotlin {
-    jvmToolchain(17)
-}
 
 android {
     namespace "design.codeux.biometric_storage"
-    compileSdk 35
+    compileSdk 36
 
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_17
-        targetCompatibility JavaVersion.VERSION_17
+        sourceCompatibility JavaVersion.VERSION_11
+        targetCompatibility JavaVersion.VERSION_11
     }
 
     kotlinOptions {
-        jvmTarget = '17'
+        jvmTarget = '11'
     }
 
     sourceSets {
@@ -64,4 +59,6 @@ dependencies {
     implementation "org.slf4j:slf4j-api:2.0.7"
     implementation "androidx.biometric:biometric:$biometric_version"
     implementation "io.github.oshai:kotlin-logging-jvm:5.0.1"
+
+    testImplementation 'junit:junit:4.13.2'
 }
