--- biometric_storage_darwin/ios/Classes/BiometricStoragePlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface BiometricStoragePlugin : NSObject<FlutterPlugin>
-@end
