--- biometric_storage_darwin/macos/Classes/BiometricStoragePlugin.swift
@@ -1,14 +1,0 @@
-import Cocoa
-import FlutterMacOS
-
-public class BiometricStoragePlugin: NSObject, FlutterPlugin {
-  public static func register(with registrar: FlutterPluginRegistrar) {
-    let channel = FlutterMethodChannel(name: "biometric_storage", binaryMessenger: registrar.messenger)
-    let instance = BiometricStoragePlugin()
-    registrar.addMethodCallDelegate(instance, channel: channel)
-  }
-
-  public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
-    result(FlutterMethodNotImplemented)
-  }
-}
