+++ biometric_storage_darwin/macos/biometric_storage_darwin.podspec
@@ -9,9 +9,9 @@ Darwin implementation for biometric_storage.
   s.license          = { :file => '../../LICENSE' }
   s.author           = { 'AuthPass' => 'support@authpass.app' }
   s.source           = { :path => '.' }
-  s.source_files     = 'Classes/**/*'
+  s.source_files     = 'biometric_storage_darwin/Sources/biometric_storage_darwin/**/*.swift'
   s.dependency 'FlutterMacOS'
-  s.platform = :osx, '10.11'
+  s.platform = :osx, '10.15'
   s.pod_target_xcconfig = { 'DEFINES_MODULE' => 'YES' }
   s.swift_version = '5.0'
 end
