+++ biometric_storage_darwin/macos/biometric_storage_darwin/Sources/biometric_storage_darwin/BiometricStorageMacOSPlugin.swift
@@ -1,13 +1,15 @@
-import Flutter
+import Cocoa
+import FlutterMacOS
 
-public class SwiftBiometricStoragePlugin: NSObject, FlutterPlugin {
+public class BiometricStorageMacOSPlugin: NSObject, FlutterPlugin {
+  
   private let impl = BiometricStorageImpl(storageError: { (code, message, details) -> Any in
     FlutterError(code: code, message: message, details: details)
   }, storageMethodNotImplemented: FlutterMethodNotImplemented)
-
+  
   public static func register(with registrar: FlutterPluginRegistrar) {
-    let channel = FlutterMethodChannel(name: "biometric_storage", binaryMessenger: registrar.messenger())
-    let instance = SwiftBiometricStoragePlugin()
+    let channel = FlutterMethodChannel(name: "biometric_storage", binaryMessenger: registrar.messenger)
+    let instance = BiometricStorageMacOSPlugin()
     registrar.addMethodCallDelegate(instance, channel: channel)
   }
