+++ biometric_storage_darwin/pubspec.yaml
@@ -5,8 +5,8 @@ homepage: https://github.com/authpass/biometric_storage/
 publish_to: 'none'
 
 environment:
-  sdk: '>=3.2.0 <4.0.0'
-  flutter: ">=3.4.0"
+  sdk: '>=3.11.0 <4.0.0'
+  flutter: ">=3.41.0"
 
 dependencies:
   flutter:
