+++ biometric_storage_windows/lib/biometric_storage_windows.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: non_constant_identifier_names
+
 import 'dart:convert';
 import 'dart:ffi';
 import 'dart:typed_data';
@@ -5,9 +7,85 @@ import 'dart:typed_data';
 import 'package:biometric_storage_platform_interface/biometric_storage_platform_interface.dart';
 import 'package:ffi/ffi.dart';
 import 'package:logging/logging.dart';
-import 'package:win32/win32.dart';
 
 final _logger = Logger('biometric_storage_windows');
+
+const _credTypeGeneric = 1;
+const _credPersistLocalMachine = 2;
+const _errorNotFound = 1168;
+
+final DynamicLibrary _advapi32 = DynamicLibrary.open('Advapi32.dll');
+final DynamicLibrary _kernel32 = DynamicLibrary.open('Kernel32.dll');
+
+final int Function(Pointer<Utf16>, int, int) _credDelete =
+    _advapi32.lookupFunction<Int32 Function(Pointer<Utf16>, Uint32, Uint32),
+        int Function(Pointer<Utf16>, int, int)>('CredDeleteW');
+
+final int Function(Pointer<Utf16>, int, int, Pointer<Pointer<_Credential>>)
+    _credRead = _advapi32.lookupFunction<
+        Int32 Function(
+          Pointer<Utf16>,
+          Uint32,
+          Uint32,
+          Pointer<Pointer<_Credential>>,
+        ),
+        int Function(
+          Pointer<Utf16>,
+          int,
+          int,
+          Pointer<Pointer<_Credential>>,
+        )>('CredReadW');
+
+final int Function(Pointer<_Credential>, int) _credWrite =
+    _advapi32.lookupFunction<Int32 Function(Pointer<_Credential>, Uint32),
+        int Function(Pointer<_Credential>, int)>('CredWriteW');
+
+final void Function(Pointer<Void>) _credFree = _advapi32
+    .lookupFunction<Void Function(Pointer<Void>), void Function(Pointer<Void>)>(
+  'CredFree',
+);
+
+final int Function() _getLastError =
+    _kernel32.lookupFunction<Uint32 Function(), int Function()>('GetLastError');
+
+final class _FileTime extends Struct {
+  @Uint32()
+  external int dwLowDateTime;
+
+  @Uint32()
+  external int dwHighDateTime;
+}
+
+final class _Credential extends Struct {
+  @Uint32()
+  external int Flags;
+
+  @Uint32()
+  external int Type;
+
+  external Pointer<Utf16> TargetName;
+
+  external Pointer<Utf16> Comment;
+
+  external _FileTime LastWritten;
+
+  @Uint32()
+  external int CredentialBlobSize;
+
+  external Pointer<Uint8> CredentialBlob;
+
+  @Uint32()
+  external int Persist;
+
+  @Uint32()
+  external int AttributeCount;
+
+  external Pointer<Void> Attributes;
+
+  external Pointer<Utf16> TargetAlias;
+
+  external Pointer<Utf16> UserName;
+}
 
 class BiometricStorageWindows extends BiometricStoragePlatform {
   static const namePrefix = 'design.codeux.authpass.';
@@ -40,13 +118,12 @@ class BiometricStorageWindows extends BiometricStoragePlatform {
     String name,
     PromptInfo promptInfo,
   ) async {
-    final namePointer =
-        PCWSTR(_storageName(name).toNativeUtf16(allocator: calloc));
+    final namePointer = _storageName(name).toNativeUtf16(allocator: calloc);
     try {
-      final result = CredDelete(namePointer, CRED_TYPE_GENERIC);
-      if (!result.value) {
-        final errorCode = result.error;
-        if (errorCode == ERROR_NOT_FOUND) {
+      final result = _credDelete(namePointer, _credTypeGeneric, 0);
+      if (result == 0) {
+        final errorCode = _getLastError();
+        if (errorCode == _errorNotFound) {
           _logger.fine('Unable to find credential of name $name');
         } else {
           _logger.warning('Error deleting credential $name: $errorCode');
@@ -65,14 +142,13 @@ class BiometricStorageWindows extends BiometricStoragePlatform {
     PromptInfo promptInfo, {
     bool forceBiometricAuthentication = false,
   }) async {
-    final credPointer = calloc<Pointer<CREDENTIAL>>();
-    final namePointer =
-        PCWSTR(_storageName(name).toNativeUtf16(allocator: calloc));
+    final credPointer = calloc<Pointer<_Credential>>();
+    final namePointer = _storageName(name).toNativeUtf16(allocator: calloc);
     try {
-      final result = CredRead(namePointer, CRED_TYPE_GENERIC, credPointer);
-      if (!result.value) {
-        final errorCode = result.error;
-        if (errorCode == ERROR_NOT_FOUND) {
+      final result = _credRead(namePointer, _credTypeGeneric, 0, credPointer);
+      if (result == 0) {
+        final errorCode = _getLastError();
+        if (errorCode == _errorNotFound) {
           _logger.fine('Unable to find credential of name $name');
         } else {
           _logger.warning('Error reading credential $name: $errorCode');
@@ -89,7 +165,7 @@ class BiometricStorageWindows extends BiometricStoragePlatform {
         cred.CredentialBlob.asTypedList(cred.CredentialBlobSize),
       );
       final value = utf8.decode(blob);
-      CredFree(credPointer.value);
+      _credFree(credPointer.value.cast<Void>());
       return value;
     } finally {
       calloc.free(credPointer);
@@ -105,27 +181,28 @@ class BiometricStorageWindows extends BiometricStoragePlatform {
     bool forceBiometricAuthentication = false,
   }) async {
     final passwordBytes = Uint8List.fromList(utf8.encode(content));
-    final blob = passwordBytes.isEmpty
-        ? nullptr
-        : passwordBytes.toNative(allocator: calloc);
-    final namePointer =
-        PWSTR(_storageName(name).toNativeUtf16(allocator: calloc));
-    final userNamePointer = PWSTR(
-      'flutter.biometric_storage'.toNativeUtf16(allocator: calloc),
+    final blob =
+        passwordBytes.isEmpty ? nullptr : calloc<Uint8>(passwordBytes.length);
+    if (blob != nullptr) {
+      blob.asTypedList(passwordBytes.length).setAll(0, passwordBytes);
+    }
+    final namePointer = _storageName(name).toNativeUtf16(allocator: calloc);
+    final userNamePointer = 'flutter.biometric_storage'.toNativeUtf16(
+      allocator: calloc,
     );
 
-    final credential = calloc<CREDENTIAL>()
-      ..ref.Type = CRED_TYPE_GENERIC
+    final credential = calloc<_Credential>()
+      ..ref.Type = _credTypeGeneric
       ..ref.TargetName = namePointer
-      ..ref.Persist = CRED_PERSIST_LOCAL_MACHINE
+      ..ref.Persist = _credPersistLocalMachine
       ..ref.UserName = userNamePointer
       ..ref.CredentialBlob = blob
       ..ref.CredentialBlobSize = passwordBytes.length;
     try {
-      final result = CredWrite(credential, 0);
-      if (!result.value) {
+      final result = _credWrite(credential, 0);
+      if (result == 0) {
         throw BiometricStorageException(
-          'Error writing credential $name: ${result.error}',
+          'Error writing credential $name: ${_getLastError()}',
         );
       }
     } finally {
